# bin/Debug/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "KYC Nintex Integration API",
    "description": "API endpoints for fetching Customer details and Customer Documents with Base64 encoded attachments for Nintex workflows.",
    "version": "v1"
  },
  "paths": {
    "/api/Customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Retrieves detailed information for a specific customer by CustomerID.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The unique Customer ID.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns customer details successfully.",
            "schema": {
              "$ref": "#/definitions/CustomerDataResponseDto"
            }
          },
          "404": {
            "description": "If no customer is found with the provided ID.",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/customers/{customerId}/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Retrieves all documents for a specific customer by CustomerID, including Base64-encoded file binary content.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "description": "The unique Customer ID.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns customer documents successfully.",
            "schema": {
              "$ref": "#/definitions/CustomerDocumentListResponseDto"
            }
          },
          "404": {
            "description": "If the customer does not exist.",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "KYCNintexApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "CustomerDataResponseDto": {
      "type": "object",
      "properties": {
        "customer": {
          "$ref": "#/definitions/CustomerResponseDto"
        }
      },
      "additionalProperties": false
    },
    "CustomerDocumentListResponseDto": {
      "type": "object",
      "properties": {
        "documents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerDocumentResponseDto"
          }
        }
      },
      "additionalProperties": false
    },
    "CustomerDocumentResponseDto": {
      "type": "object",
      "properties": {
        "documentID": {
          "format": "int32",
          "type": "integer"
        },
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "documentType": {
          "type": "string"
        },
        "fileName": {
          "type": "string"
        },
        "issueDate": {
          "type": "string"
        },
        "expiryDate": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "fileContentBase64": {
          "description": "Base64 encoded file content binary wrapped inside JSON for Nintex buffer processing.",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CustomerResponseDto": {
      "type": "object",
      "properties": {
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "fullName": {
          "type": "string"
        },
        "dateOfBirth": {
          "type": "string"
        },
        "permanentAddress": {
          "type": "string"
        },
        "mailingAddress": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "branch": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "dateOfInitiation": {
          "format": "date-time",
          "type": "string"
        },
        "lastUpdatedDate": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    }
  }
}